+++ src/modules/hip/kernel/contrast.hpp
@@ -1,5 +1,5 @@
 #include <hip/hip_runtime.h>
-#include "hip/rpp_hip_common.hpp"
+#include "rpp_hip_common.hpp"
 
 __device__ void contrast_hip_compute(uchar *srcPtr, d_float8 *pix_f8, d_float8 *contrastParams_f8)
 {
