+++ src/modules/hip/kernel/noise_salt_and_pepper.hpp
@@ -1,6 +1,6 @@
 #include <hip/hip_runtime.h>
-#include "hip/rpp_hip_common.hpp"
-#include "func_specific/rng_seed_stream.hpp"
+#include "rpp_hip_common.hpp"
+#include "rng_seed_stream.hpp"
 
 __device__ void salt_and_pepper_noise_1_hip_compute(float *pix, float noiseProbability, float saltProbability, float salt, float pepper, float randomNumberFloat)
 {
